# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SurveySubmissionHoursPerWeek =
  (typeof SurveySubmissionHoursPerWeek)[keyof typeof SurveySubmissionHoursPerWeek];

export const SurveySubmissionHoursPerWeek = {
  "1-2_hours": "1-2 hours",
  "2-4_hours": "2-4 hours",
  "4-6_hours": "4-6 hours",
  "8-10_hours": "8-10 hours",
  more_than_10_hours: "more than 10 hours",
} as const;

export interface SurveySubmission {
  /**
   * @minimum 18
   * @maximum 24
   */
  age: number;
  /** @minLength 1 */
  major: string;
  hoursPerWeek: SurveySubmissionHoursPerWeek;
  /** @minLength 1 */
  freeTimeHobbies: string;
}

export interface SurveySubmitResponse {
  id: number;
  message: string;
}

export interface AgeBucket {
  age: number;
  count: number;
}

export interface MajorBucket {
  major: string;
  count: number;
}

export interface HoursBucket {
  hoursPerWeek: string;
  count: number;
}

export interface HobbyBucket {
  hobby: string;
  count: number;
}

export interface SurveyResults {
  totalResponses: number;
  ageDistribution: AgeBucket[];
  majorDistribution: MajorBucket[];
  hoursDistribution: HoursBucket[];
  hobbyDistribution: HobbyBucket[];
}

export interface ErrorResponse {
  error: string;
  details?: string;
}