# lib/api-zod/src/generated/types/surveyResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */
import type { AgeBucket } from "./ageBucket";
import type { HobbyBucket } from "./hobbyBucket";
import type { HoursBucket } from "./hoursBucket";
import type { MajorBucket } from "./majorBucket";

export interface SurveyResults {
  totalResponses: number;
  ageDistribution: AgeBucket[];
  majorDistribution: MajorBucket[];
  hoursDistribution: HoursBucket[];
  hobbyDistribution: HobbyBucket[];
}