# lib/api-zod/src/generated/types/surveySubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveySubmissionHoursPerWeek } from "./surveySubmissionHoursPerWeek";

export interface SurveySubmission {
  /**
   * @minimum 18
   * @maximum 24
   */
  age: number;
  /** @minLength 1 */
  major: string;
  hoursPerWeek: SurveySubmissionHoursPerWeek;
  /** @minLength 1 */
  freeTimeHobbies: string;
}