# lib/api-zod/src/generated/types/surveySubmitResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Free Time Survey API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveySubmitResponse {
  id: number;
  message: string;
}